Fixed some Typos / null checks
Fixed misspellings in error messages.
var changed to let and const.
Added null and undefined checks where needed.
Text and tabs adapted to eslint rules.

TO DO:
_Ke could be undefined, a correct error message must still be written.
_Kd could be undefined, a proper error message has yet to be written.

problems:
"AES" does not have the "key" property.
"this" in "Object.defineProperty" is implicitly of type "any" because there is no type annotation.

Error in the lower area:
"define" is not defined.
"root" is not defined.